Add files via upload
The old pull request was to see what I had deleted, the main branch is actually the one that was working, so this is an update beyond the Main branch. What is working is myStories section, Hack-or-Snooze link, Login. What is not working is: myFavorites, & can't get the Star-click to work. This is holding up the deleting process.